+++ CHANGELOG.md
@@ -15,6 +15,7 @@
   - Internal `setValue` has argument `reason` now
   - Event `$change` and callback `$onChange` has prop `detail: SetValueReason` now
   - Added `$defaults.storage` and `$options.skey` to store/get value in different storages `local/session/url`
+  - Extended custom validation to `(value, control) => false | string`. Was `(value) => false | string`
 - **Glolbal**. Added JSDoc comments for `jsx/tsx` files (previously hovering on `<wup-text>` showed nothing)
 
 ## 0.6.0 (May 19, 2023)